nytprofhtml:  Make fragment target visible.
With floatThead.js, fragment targets are hidden by the floating table
headers.  This patch fixes this.  The uppermost visible position of
the target is computed and the window scrolled accordingly.

Also, if the user changes the fragment in the browser's address field,
make the new fragment target visible again.  This second part of the
fix might not work in old browsers.  The browser has to provide the
'window.hashchange' function.