Fix for the double opening of the web browser
When [Browser::Open](https://metacpan.org/pod/Browser::Open) is available and it opens a browser, it returns the result of `system()` (see [source code](https://metacpan.org/source/CFRANKS/Browser-Open-0.04/lib/Browser/Open.pm#L72)), which is the exit code of the process. On success, the exit code is zero, so we now check this and return without trying to open the browser again.

Previously, the browser was opening twice when the `--open` flag was provided.