t/test62-subcaller1-b.rdt: start to adapt
We copy-and-paste in new expected data where we can.

Though these revisions do not reduce the number of unit test failures
outstanding, they do reduce the line count of error output from 733
lines to 573 lines.  The only lines which have not been accounted for
are those including an absolute path to bleadperl's strict.pm file (more
specifically, to 2 coderefs therein).

  strict::__ANON__[/home/usename/blead/lib/perl5/5.33.9/strict.pm:31]

There are 64 such lines, which means each of 16
failing unit tests is associated with 4 unaccounted for lines of output.