Workaround for Devel::Cover oddity
When running 'cover -report=text', we're currently getting a 'Redundant
argument in printf' warning.  The statement is being treated as "and 2
conditions."  Experimentation shows we can suppress this
warning in Data.pm if we use a regular 'foreach' syntax.  The statement
is no longer treated by D::C as a condition.