trace_level() is tricky
Large parts of the code base are not exercised by the test suite unless
'harness' is preceded by 'export NYTPROF="trace=N"' (where N is positive
integer).  But running with trace on (a) prints enormous amount of data
and (b) can interact with "normal" dumping and other settings to
NYTPROF.  So certain unit tests must be skipped.