Remove redundant 'if' condition
In the definition of D-N-Data::collapse_evals_in(), the following
condition appeared twice:

    if (trace_level() >= 2) {

The second invocation occurred within the block of the first invocation.
Consequently, the second invocation always evaluated to true; the false
condition (trace_level() < 2) was never reached within the second,
interior block.  Coverage analysis demonstrated this:  the false part of
the branch was never exercised.

Hence, we can remove the second invocation entirely and allow the code
within to run as part of the block of the first invocation.  Running
'make test' with trace_level set to 5 had all tests continue to pass
with no meaningful change in dumped output.