Revert "Update flamegraph.pl"
This reverts commit 65ccbf3eac29bca9086fe06e9e8ebe07fb5bc171.

bin/flamegraph.pl was updated after the 6.06 release of the distribution
but before the current co-maintainer began work.  It will take some time
to assess the before/after functioning of flamegraph due to this commit.
Right now it's more urgent to get 6.07 out the door, so we'll defer
inclusion of revisions to flamegraph until a later release.