Move unused code out of the way
One function long present in NYTProf.xs was never used, but its lack of
use generates a build-time warning -- the last such warning currently
observed.  Move it to a file in the devstuff/ directory so that we can
access it in the future should we need to do so.

Similarly, two functions in Devel::NYTProf::Data were never used in
lib/, bin/ or t/, so let's similarly move them out of the way.