Don't run these tests on longdouble builds
As described in https://github.com/timbunce/devel-nytprof/issues/168,
the profile data used for testing in these 4 files is not appropriate
for builds with '-Duselongdouble'.  We get an error like this:

    Profile data created by incompatible perl config (NV size 8
    but ours is 16)

Until such time as we can provide better testing data, let's skip these
tests on longdouble builds.  (The point of these tests is simply to
extend code coverage; we don't need to run on longdouble builds to
achieve that objective.)