completely remove floatthead
Removed all code that initializes the plugin (doing this under the assumption that https://github.com/timbunce/devel-nytprof/pull/194 removes the js file from this project).
Also removed `show_fragment_target()` because it would no longer work without floatthead present. 

This fixes #192 , or at least fixes the fix that was made in it.

The proper way to fix #192 would be to take a look at the `show_fragment_target()` function and figure out why its setting the incorrect scrollTop there when you click on links. I cannot do this without seeing the generated code. 

Either way, someone needs to test this change before merging it, because I have not done so.