Test only on newer perls
This test file is incurring failures on certain CPANtesters rigs when
tested on Perl production releases from 5.8.9/5.10 through 5.32. See,
for example:

  http://www.cpantesters.org/cpan/report/7b0f4d16-6634-11ee-9717-aa48143b15f3

The failures are not consistent across different CPANtesters rigs and
would require access to particular machines in order to analyze further
(e.g., by bisection).

The unit tests fail because the output data no longer matches the data
calculated many years ago, but it's not self-evident that these
differences are meaningful from a day-to-day profiling perspective.

If you wish to see how this test file performs on 5.32 and earlier,
please set environmental variable NYTPROF_AUTHOR_TESTING to 1.