Update link to Flame Graph info in nytprofhtml output
While investigating the profiling output on a project I'm working on I
happened to notice that the link to Brendan Gregg's Flame Graph page was
broken.  It seems that the new page is now on www.brendangregg.com (as
implemented in this change).

The last snapshot on the internet archive to still point to the
dtrace.org address is from November 2013:
https://web.archive.org/web/20131110195734/http://dtrace.org/blogs/brendan/2011/12/16/flame-graphs/
and the following snapshot in January 2014 points to the new address
https://web.archive.org/web/20140121022436/http://www.brendangregg.com/FlameGraphs/cpuflamegraphs.html,
which then seems to have morphed into the link I've used here.  Thus,
I'm fairly sure that the link I've used to fix the broken one is
correct.